Add policy and overflow, mode and type arguments as labels on queues
Labels should not have values with high cardinalities, as each label
generates a new time series, so keys such as exclusive_consumer_tag
should not be stored in prometheus. We could possibly argue that policy
is one of those, but it might be worth to try it out as it could be a
very useful metric.