Add ezs make target that builds .ez files for all deps & plugin
The resulting files will be placed in ./tmp, e.g.

    ls -l tmp/
    total 992
    -rw-r--r--  1 gerhard  staff   12548 20 Feb 18:02 accept-0.3.4.ez
    -rw-r--r--  1 gerhard  staff  211232 20 Feb 18:02 prometheus-4.2.2.ez
    -rw-r--r--  1 gerhard  staff   14053 20 Feb 18:02 prometheus_cowboy-0.1.7.ez
    -rw-r--r--  1 gerhard  staff   21638 20 Feb 18:02 prometheus_httpd-2.1.10.ez
    -rw-r--r--  1 gerhard  staff   10708 20 Feb 18:02 prometheus_process_collector-1.4.0.ez
    -rw-r--r--  1 gerhard  staff  223943 20 Feb 18:02 prometheus_rabbitmq_exporter-3.7.20190220.ez

I've changed the version of this plugin to use the date for the patch
part, feel free to change back if you disagree.

Notice that I've deleted the prometheus_rabbitmq_exporter.app.src, this
is now generated by erlang.mk and dropped in
ebin/prometheus_rabbitmq_exporter.app - does it look correct
@deadtrickster?

Want to integrate with Docker & testing before merging.

This started as deadtrickster/prometheus_rabbitmq_exporter#63, see the
discussion for more context.