Refactor Docker make targets to make use of the new ezs
The resulting Docker images are only tagged with the RabbitMQ minor
version, since prometheus_rabbitmq_exporter is expected to work with the
latest stable RabbitMQ, regardless which that is at time of building.
When these images are built, since we are using the `--pull` flag, we
will always get the latest FROM image version at time of building.
Without any automation, manually updating this version to the latest
upstream feels inefficient and quiet boring - that would explain the
version gaps in
https://hub.docker.com/r/deadtrickster/rabbitmq_prometheus/tags, so I'm
thinking to not use the exact RabbitMQ version in the tag for this
image. I am on-board with using the PROJECT_VERSION (e.g. 3.7.2.4) if
you think that is better @deadtrickster.

Deleted the pure variant, it's just a duplicate of Dockerfile since #54,
prometheus_process_collector ends up in both image variants.