Add make target that allows building ezs in a specific OTP context
This was necessary to test that the fixes for #75 work in RabbitMQ
v3.7.9 & OTP v21.2.5, the combo that ships in Docker image
rabbitmq:3.7.9. This is the first RabbitMQ Docker image that supports
prometheus_rabbitmq_exporter (compiled with OTP v21.0.2). Docker image
rabbitmq:3.7.8 ships with OTP v20.3.8.5

Notice that I disable prometheus_process_collector in the Docker build
context. This fails to build due to Docker image erlang:21.2.5 missing
native libs. Didn't want to shave this yak as well.